[benchmark] Extract common binary utils out of RuBen.
Summary:
This code change introduce no new features, but only move things around so that
in future throughput test code and linear search binary can share the same utils.

Test Plan:
Make sure everything works as usual.
