+++ benchmark/src/lib.rs
@@ -17,6 +17,7 @@ use rand::Rng;
 use std::{collections::HashMap, convert::TryInto, sync::Arc, thread, time};
 use types::{account_address::AccountAddress, account_config::association_address};
 
+pub mod bin_utils;
 pub mod grpc_helpers;
 pub mod ruben_opt;
 pub mod submit_rate;
